% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{limite}
\alias{limite}
\title{Construct a Limite language model}
\usage{
limite(config = limite_config())
}
\arguments{
\item{config}{A configuration produced by \code{\link[=limite_config]{limite_config()}} or loaded from a
checkpoint.}
}
\value{
A \code{torch::nn_module} implementing \code{LimiteForCausalLM}.
}
\description{
Construct a Limite language model
}
