% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{limite_tokenizer}
\alias{limite_tokenizer}
\title{Load the Limite tokenizer}
\usage{
limite_tokenizer(
  identifier = "paradigma-inc/limite-1b-violetto",
  revision = "main",
  local_files_only = FALSE
)
}
\arguments{
\item{identifier}{Model identifier or local path.}

\item{revision}{Hugging Face revision.}

\item{local_files_only}{If \code{TRUE}, do not download missing Hub files.}
}
\value{
A \code{tok_tokenizer}.
}
\description{
Load the Limite tokenizer
}
